Logging, proper database initialization and a first version of the authentication endpoint
