+++ .gitignore
@@ -1,8 +1,5 @@
 .DS_Store*
 
-# VSCode
-.vscode
-
 # Logs
 logs
 error.log
@@ -17,6 +14,7 @@ pids
 *.pid
 *.seed
 *.pid.lock
+*.db
 
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
