+++ advanced/challenge-1/backend/app/db/database.py
@@ -2,7 +2,19 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import declarative_base, sessionmaker
 
 engine = create_engine(
-    'sqlite+pysqlite:///:memory:?check_same_thread=true', echo=True, future=True)
+    'sqlite+pysqlite:///./app.db?check_same_thread=true', echo=False, future=True)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 Base = declarative_base()
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+def get_conn():
+    return next(get_db())
