+++ advanced/challenge-1/backend/logging.conf
@@ -1,0 +1,46 @@
+[loggers]
+keys=root,db,null
+
+[handlers]
+keys=root_handler,db_handler,null_handler
+
+[formatters]
+keys=root_formatter
+
+[logger_root]
+level=INFO
+handlers=root_handler
+
+[logger_db]
+level=INFO
+handlers=db_handler
+qualname=db
+propagate=0
+
+[logger_null]
+level=INFO
+handlers=null_handler
+qualname=watchfiles
+propagate=0
+
+[handler_root_handler]
+class=FileHandler
+args=('app.log','a')
+level=INFO
+formatter=root_formatter
+
+[handler_db_handler]
+class=FileHandler
+args=('app.log','a')
+level=INFO
+formatter=root_formatter
+
+[handler_null_handler]
+class=StreamHandler
+args=(sys.stdout,)
+level=INFO
+formatter=root_formatter
+
+[formatter_root_formatter]
+format=%(asctime)s,%(msecs)03d %(levelname)-8s %(name)s:l%(lineno)-4d %(message)s
+
