+++ advanced/challenge-1/frontend/.env
@@ -1,0 +1,1 @@
+NODE_ENV = "development"
